export and copy report happens in main window
