+++ openpype/tools/publisher/window.py
@@ -1,3 +1,6 @@
+import os
+import json
+import time
 import collections
 import copy
 from qtpy import QtWidgets, QtCore, QtGui
@@ -15,6 +18,7 @@ from openpype.tools.utils import (
 
 from .constants import ResetKeySequence
 from .publish_report_viewer import PublishReportViewerWidget
+from .control import CardMessageTypes
 from .control_qt import QtPublisherController
 from .widgets import (
     OverviewWidget,
@@ -313,6 +317,13 @@ class PublisherWindow(QtWidgets.QDialog):
         controller.event_system.add_callback(
             "convertors.find.failed", self._on_convertor_error
         )
+        controller.event_system.add_callback(
+            "export_report.request", self._export_report
+        )
+        controller.event_system.add_callback(
+            "copy_report.request", self._copy_report
+        )
+
 
         # Store extra header widget for TrayPublisher
         # - can be used to add additional widgets to header between context
@@ -943,6 +954,46 @@ class PublisherWindow(QtWidgets.QDialog):
             widget_x = widget_geo.left() + (widget_geo.width() * 0.5)
             under_mouse = widget_x < global_pos.x()
         self._create_overlay_button.set_under_mouse(under_mouse)
+
+    def _copy_report(self):
+        logs = self._controller.get_publish_report()
+        logs_string = json.dumps(logs, indent=4)
+
+        mime_data = QtCore.QMimeData()
+        mime_data.setText(logs_string)
+        QtWidgets.QApplication.instance().clipboard().setMimeData(
+            mime_data
+        )
+        self._controller.emit_card_message(
+            "Report added to clipboard",
+            CardMessageTypes.info)
+
+    def _export_report(self):
+        default_filename = "publish-report-{}".format(
+            time.strftime("%y%m%d-%H-%M")
+        )
+        default_filepath = os.path.join(
+            os.path.expanduser("~"),
+            default_filename
+        )
+        new_filepath, ext = QtWidgets.QFileDialog.getSaveFileName(
+            self, "Save report", default_filepath, ".json"
+        )
+        if not ext or not new_filepath:
+            return
+
+        logs = self._controller.get_publish_report()
+        full_path = new_filepath + ext
+        dir_path = os.path.dirname(full_path)
+        if not os.path.exists(dir_path):
+            os.makedirs(dir_path)
+
+        with open(full_path, "w") as file_stream:
+            json.dump(logs, file_stream)
+
+        self._controller.emit_card_message(
+            "Report saved",
+            CardMessageTypes.info)
 
 
 class ErrorsMessageBox(ErrorMessageBox):
